formatter/refactor - split out dedicated module from utils
- no feature change refactor
