+++ lib/command.changelog.js
@@ -1,5 +1,5 @@
 const { options, baseHandler } = require('./common')
-const { formatChangelog: formatter } = require('./utils')
+const { formatChangelog: formatter } = require('./formatter')
 
 
 const command = 'changelog'
