+++ lib/command.notes.js
@@ -1,5 +1,5 @@
 const { options, baseHandler } = require('./common')
-const { formatNotes: formatter } = require('./utils')
+const { formatNotes: formatter } = require('./formatter')
 
 
 const command = 'notes'
